[refactor] Update api for masks and reimplement logic
